backend/auth_service/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings
python-dotenv
email-validator
pyotp
pytest==7.4.4
httpx==0.26.0
